% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{captions}
\alias{captions}
\alias{CAPTION_COE}
\alias{CAPTION_COE_SC}
\alias{CAPTION_TORONTO}
\title{Standard caption strings for plots and tables}
\usage{
CAPTION_COE

CAPTION_COE_SC

CAPTION_TORONTO
}
\description{
Standard caption strings for plots and tables
}
