% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{EDMONTON_RESIDENTIAL_BUILDING_TYPES}
\alias{EDMONTON_RESIDENTIAL_BUILDING_TYPES}
\title{Edmonton residential building type codes}
\usage{
EDMONTON_RESIDENTIAL_BUILDING_TYPES
}
\description{
Character vector of building type strings from the Edmonton Open Data
building permits shapefile that correspond to residential uses.
}
