% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map_dark}
\alias{theme_map_dark}
\title{Dark map theme}
\usage{
theme_map_dark()
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
Convenience wrapper combining \code{\link[=theme_map]{theme_map()}} on top of \code{\link[=theme_jd]{theme_jd()}}.
Gives a dark background with no axes and the magma colour palette.
}
