% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{captions}
\alias{captions}
\alias{CAPTION_COE}
\alias{CAPTION_COE_SC}
\alias{CAPTION_TORONTO}
\title{Standard caption strings for plots and tables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
CAPTION_COE

CAPTION_COE_SC

CAPTION_TORONTO
}
\description{
Standard caption strings for plots and tables
}
\keyword{datasets}
