% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EDMONTON_RESIDENTIAL_BUILDING_TYPES}
\alias{EDMONTON_RESIDENTIAL_BUILDING_TYPES}
\title{Edmonton residential building type codes}
\format{
An object of class \code{character} of length 15.
}
\usage{
EDMONTON_RESIDENTIAL_BUILDING_TYPES
}
\description{
Character vector of building type strings from the Edmonton Open Data
building permits shapefile that correspond to residential uses.
}
\keyword{datasets}
