% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{layers_map_base}
\alias{layers_map_base}
\title{Map base layers (water + roads)}
\usage{
layers_map_base(roads_type = c("major", "all"))
}
\arguments{
\item{roads_type}{\code{"major"} (default) to show only highways and major roads,
or \code{"all"} to show all roads at reduced opacity.}
}
\value{
A list of two ggplot2 layers.
}
\description{
Returns a list of ggplot2 layers adding water and road context to a map.
Requires the \code{mountainmathHelpers} package (listed in \code{Suggests}).
}
\examples{
\dontrun{
ggplot(data) +
  layers_map_base() +
  geom_sf(aes(colour = units_added)) +
  theme_map()
}
}
